requirements.txt
aiohttp
sanic==0.7.0
sanic-oauth
requests
sanic_session
jsonrpcclient
jsonrpcserver==3.5.6
leveldb
secp256k1
PyJWT